Move core::alloc::CollectionAllocErr to alloc::collections
